refactor: MapAction을 별도 파일로 분리 및 상세 주석 추가
- MapViewRepresentable.swift에서 MapAction 열거형을 MapViewAction.swift로 분리
- 각 MapAction case에 대한 상세한 용도와 매개변수 설명 주석 추가
- 코드 구조 개선을 통한 가독성 및 유지보수성 향상
